Working on texture

+++ Ilumination/Color.cpp
@@ -1,7 +1,7 @@
   #include "Color.h"
 
   Color::Color(){}
-  Color::Color(double red, double green, double blue){
+  Color::Color(uint8_t red, uint8_t green, uint8_t blue){
     this->red = red;// & 0xFF000000;
     this->green = green;// & 0xFF000000;
     this->blue = blue;// & 0xFF000000;
